\name{chr21genes}
     \docType{data}
     \alias{chr21genes}
     \title{Positions of Genes on Human Chromosome 21}
     \description{
	Annotation of chromosome 21 genes from RefSeq in June 2010.
     }
     \usage{chr21genes}
     \format{A data frame.}
     \source{UCSC Genome Browser tables.}
     \keyword{datasets}

