\name{laneCounts}
\alias{laneCounts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to count the total number of ChIP-seq reads in each lane.
}
\description{
A function to count the total number of ChIP-seq reads in each lane.}
\usage{
laneCounts(cs)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{cs}{A \code{GenomeDataList}.
}
}
\value{
	An \code{integer vector}.
}
\author{Aaron Statham}
\examples{
#See the manual
}
